fix(factory): split proof profile modules under maintainability caps
Extract OpenClaw probe helpers and shared verify finalization so factory
proof claim scripts and profile resolution pass hard function/file limits.
