+++ lib/task-platform/factory-proof-profile.js
@@ -1,11 +1,15 @@
 'use strict';
 
+const fs = require('node:fs');
 const path = require('node:path');
-const http = require('node:http');
-const https = require('node:https');
+const {
+  DEFAULT_LIVE_OPENCLAW_URL,
+  DEFAULT_PROBE_TIMEOUT_MS,
+  resolveOpenClawBaseUrl,
+  probeOpenClawGateway,
+  readArg: readProbeArg,
+} = require('./factory-proof-probe');
 
-const DEFAULT_LIVE_OPENCLAW_URL = 'http://127.0.0.1:18789';
-const DEFAULT_PROBE_TIMEOUT_MS = 3000;
 const OPENCLAW_SPECIALIST_RUNNER = `node ${path.resolve(process.cwd(), 'scripts/openclaw-specialist-runner.js')}`;
 const FIXTURE_SPECIALIST_RUNNER = `node ${path.resolve(process.cwd(), 'tests/fixtures/specialist-runtime-runner.js')}`;
 
@@ -20,6 +24,13 @@ const FACTORY_PROOF_ERROR_CODES = Object.freeze({
 const FIXTURE_WARNING =
   'FACTORY_PROOF_PROFILE=fixture: results are not valid for operator-trusted factory claims.';
 
+const GATEWAY_REMEDIATION = [
+  'Start the OpenClaw gateway (local default http://127.0.0.1:18789).',
+  'Pass --openclaw-url or set OPENCLAW_BASE_URL.',
+  'Ensure the audit API process also has OPENCLAW_BASE_URL, FF_REAL_SPECIALIST_DELEGATION=true, and SPECIALIST_DELEGATION_RUNNER=node scripts/openclaw-specialist-runner.js.',
+  'For non-claim local smoke only: --allow-fixture-delegation or FACTORY_PROOF_PROFILE=fixture.',
+].join(' ');
+
 function parseBoolean(value, fallback = false) {
   if (value == null || value === '') return fallback;
   return ['1', 'true', 'yes', 'on'].includes(String(value).trim().toLowerCase());
@@ -30,9 +41,7 @@ function hasFlag(argv = process.argv, name) {
 }
 
 function readArg(argv = process.argv, name, fallback = '') {
-  if (!Array.isArray(argv)) return fallback;
-  const index = argv.indexOf(name);
-  return index === -1 || index === argv.length - 1 ? fallback : argv[index + 1];
+  return readProbeArg(argv, name, fallback);
 }
 
 function createFactoryProofError(code, message, details = {}) {
@@ -49,150 +58,30 @@ function isFixtureDelegationRunner(value) {
 function isFixtureSessionId(sessionId) {
   const value = String(sessionId || '').trim();
   if (!value) return false;
-  return (
-    value.startsWith('runtime-session-')
+  return value.startsWith('runtime-session-')
     || value.startsWith('fixture-session-')
-    || /^fixture[-_]/i.test(value)
-  );
+    || /^fixture[-_]/i.test(value);
 }
 
 function isFixtureOwnership(ownership = {}) {
-  const runtime = String(ownership.runtime || ownership.mode || '').toLowerCase();
-  return runtime.includes('fixture');
+  return String(ownership.runtime || ownership.mode || '').toLowerCase().includes('fixture');
 }
 
 function isProductionLikeProof(options = {}, env = process.env) {
   const profile = String(options.proofProfile || env.FACTORY_PROOF_PROFILE || '').trim().toLowerCase();
   if (profile === 'live' || profile === 'production-like' || profile === 'fail-closed') return true;
-  return (
-    options.requireRealEvidence === true
+  return options.requireRealEvidence === true
     || options.collectRealEvidence === true
     || options.agentDrivenPhases === true
     || parseBoolean(env.FF_GOLDEN_PATH_REQUIRE_REAL_EVIDENCE, false)
     || parseBoolean(env.FF_GOLDEN_PATH_COLLECT_REAL_EVIDENCE, false)
     || parseBoolean(env.FF_REAL_SPECIALIST_DELEGATION, false)
-    || String(env.FACTORY_PROOF_PROFILE || '').toLowerCase() === 'live'
-  );
-}
-
-function resolveOpenClawBaseUrl({
-  argv = process.argv,
-  env = process.env,
-  openclawUrl = '',
-  preferDefaultLive = false,
-} = {}) {
-  const fromArg = openclawUrl || readArg(argv, '--openclaw-url', '');
-  const fromEnv = env.OPENCLAW_BASE_URL || '';
-  if (fromArg) return String(fromArg).trim();
-  if (fromEnv) return String(fromEnv).trim();
-  if (preferDefaultLive) return DEFAULT_LIVE_OPENCLAW_URL;
-  return '';
-}
-
-function requestProbe(url, timeoutMs, fetchImpl) {
-  if (typeof fetchImpl === 'function') {
-    const controller = typeof AbortController !== 'undefined' ? new AbortController() : null;
-    const timer = controller
-      ? setTimeout(() => controller.abort(), timeoutMs)
-      : null;
-    return fetchImpl(url, {
-      method: 'GET',
-      signal: controller?.signal,
-      headers: { accept: 'application/json, text/plain, */*' },
-    }).finally(() => {
-      if (timer) clearTimeout(timer);
-    }).then((response) => ({
-      ok: Boolean(response?.ok || (response?.status >= 200 && response?.status < 500)),
-      status: response?.status,
-    }));
-  }
-
-  return new Promise((resolve, reject) => {
-    const parsed = new URL(url);
-    const lib = parsed.protocol === 'https:' ? https : http;
-    const req = lib.request(
-      {
-        protocol: parsed.protocol,
-        hostname: parsed.hostname,
-        port: parsed.port || (parsed.protocol === 'https:' ? 443 : 80),
-        path: `${parsed.pathname || '/'}${parsed.search || ''}`,
-        method: 'GET',
-        timeout: timeoutMs,
-        headers: { accept: 'application/json, text/plain, */*' },
-      },
-      (res) => {
-        res.resume();
-        resolve({
-          ok: res.statusCode >= 200 && res.statusCode < 500,
-          status: res.statusCode,
-        });
-      },
-    );
-    req.on('timeout', () => {
-      req.destroy(createFactoryProofError(
-        FACTORY_PROOF_ERROR_CODES.GATEWAY_UNAVAILABLE,
-        `OpenClaw probe timed out after ${timeoutMs}ms`,
-      ));
-    });
-    req.on('error', reject);
-    req.end();
-  });
-}
-
-async function probeOpenClawGateway(options = {}) {
-  const baseUrl = String(options.baseUrl || '').trim().replace(/\/$/, '');
-  const timeoutMs = Number(options.timeoutMs || process.env.OPENCLAW_PROBE_TIMEOUT_MS || DEFAULT_PROBE_TIMEOUT_MS);
-  const started = Date.now();
-  if (!baseUrl) {
-    return {
-      available: false,
-      baseUrl: '',
-      latencyMs: 0,
-      errorCode: FACTORY_PROOF_ERROR_CODES.GATEWAY_UNAVAILABLE,
-      errorMessage: 'No OpenClaw base URL configured for probe',
-    };
-  }
-
-  const candidates = [
-    baseUrl,
-    `${baseUrl}/health`,
-    `${baseUrl}/v1/health`,
-    `${baseUrl}/api/health`,
-  ];
-
-  let lastError = null;
-  for (const candidate of candidates) {
-    try {
-      const result = await requestProbe(candidate, timeoutMs, options.fetchImpl);
-      if (result.ok) {
-        return {
-          available: true,
-          baseUrl,
-          probedUrl: candidate,
-          status: result.status,
-          latencyMs: Date.now() - started,
-        };
-      }
-      lastError = `HTTP ${result.status}`;
-    } catch (error) {
-      lastError = error?.message || String(error);
-    }
-  }
-
-  return {
-    available: false,
-    baseUrl,
-    latencyMs: Date.now() - started,
-    errorCode: FACTORY_PROOF_ERROR_CODES.GATEWAY_UNAVAILABLE,
-    errorMessage: lastError || 'OpenClaw gateway probe failed',
-  };
+    || String(env.FACTORY_PROOF_PROFILE || '').toLowerCase() === 'live';
 }
 
 function resolveExplicitProfile({ argv = process.argv, env = process.env } = {}) {
   const envProfile = String(env.FACTORY_PROOF_PROFILE || '').trim().toLowerCase();
-  if (envProfile === 'fixture' || envProfile === 'live' || envProfile === 'fail-closed') {
-    return envProfile;
-  }
+  if (envProfile === 'fixture' || envProfile === 'live' || envProfile === 'fail-closed') return envProfile;
   if (
     hasFlag(argv, '--allow-fixture-delegation')
     || hasFlag(argv, '--fixture-delegation')
@@ -200,16 +89,42 @@ function resolveExplicitProfile({ argv = process.argv, env = process.env } = {})
   ) {
     return 'fixture';
   }
-  if (hasFlag(argv, '--live-openclaw') || hasFlag(argv, '--require-live-openclaw')) {
-    return 'live';
-  }
+  if (hasFlag(argv, '--live-openclaw') || hasFlag(argv, '--require-live-openclaw')) return 'live';
   return null;
 }
 
-/**
- * Resolve the factory proof profile for claim / verify entrypoints.
- * Primary default: live when gateway available, otherwise fail closed.
- */
+function buildFixtureProof(baseUrl, probe = null) {
+  return {
+    profile: 'fixture',
+    fixtureAllowed: true,
+    fixtureDelegation: true,
+    openclawBaseUrl: baseUrl || null,
+    runner: FIXTURE_SPECIALIST_RUNNER,
+    probe,
+    warning: FIXTURE_WARNING,
+  };
+}
+
+function buildLiveProof(baseUrl, probe) {
+  return {
+    profile: 'live',
+    fixtureAllowed: false,
+    fixtureDelegation: false,
+    openclawBaseUrl: probe.baseUrl || baseUrl,
+    runner: OPENCLAW_SPECIALIST_RUNNER,
+    probe,
+    warning: null,
+  };
+}
+
+function throwGatewayUnavailable(baseUrl, probe, explicit) {
+  throw createFactoryProofError(
+    FACTORY_PROOF_ERROR_CODES.GATEWAY_UNAVAILABLE,
+    `Factory proof gateway unavailable at ${baseUrl || '(unset)'}: ${probe.errorMessage || 'probe failed'}. ${GATEWAY_REMEDIATION}`,
+    { probe, openclawBaseUrl: baseUrl || null, profile: explicit || 'live' },
+  );
+}
+
 async function resolveFactoryProofProfile(options = {}) {
   const argv = options.argv || process.argv;
   const env = options.env || process.env;
@@ -222,73 +137,34 @@ async function resolveFactoryProofProfile(options = {}) {
     preferDefaultLive,
   });
 
-  if (explicit === 'fixture') {
-    return {
-      profile: 'fixture',
-      fixtureAllowed: true,
-      fixtureDelegation: true,
-      openclawBaseUrl: baseUrl || null,
-      runner: FIXTURE_SPECIALIST_RUNNER,
-      probe: options.probe || null,
-      warning: FIXTURE_WARNING,
-    };
-  }
+  if (explicit === 'fixture') return buildFixtureProof(baseUrl, options.probe || null);
 
   const probe = options.probe || await probeOpenClawGateway({
     baseUrl,
     timeoutMs: options.timeoutMs,
     fetchImpl: options.fetchImpl,
   });
-
-  if (probe.available) {
-    return {
-      profile: 'live',
-      fixtureAllowed: false,
-      fixtureDelegation: false,
-      openclawBaseUrl: probe.baseUrl || baseUrl,
-      runner: OPENCLAW_SPECIALIST_RUNNER,
-      probe,
-      warning: null,
-    };
-  }
-
-  // live or default claim path: fail closed when gateway is down
-  const remediation = [
-    'Start the OpenClaw gateway (local default http://127.0.0.1:18789).',
-    'Pass --openclaw-url or set OPENCLAW_BASE_URL.',
-    'Ensure the audit API process also has OPENCLAW_BASE_URL, FF_REAL_SPECIALIST_DELEGATION=true, and SPECIALIST_DELEGATION_RUNNER=node scripts/openclaw-specialist-runner.js.',
-    'For non-claim local smoke only: --allow-fixture-delegation or FACTORY_PROOF_PROFILE=fixture.',
-  ].join(' ');
-
-  throw createFactoryProofError(
-    FACTORY_PROOF_ERROR_CODES.GATEWAY_UNAVAILABLE,
-    `Factory proof gateway unavailable at ${baseUrl || '(unset)'}: ${probe.errorMessage || 'probe failed'}. ${remediation}`,
-    { probe, openclawBaseUrl: baseUrl || null, profile: explicit || 'live' },
-  );
+  if (probe.available) return buildLiveProof(baseUrl, probe);
+  throwGatewayUnavailable(baseUrl, probe, explicit);
 }
 
 function applyFactoryProofProfileToEnv(proof, env = process.env) {
-  const next = env;
-  next.FACTORY_PROOF_PROFILE = proof.profile;
-  next.FACTORY_USE_FIXTURE_DELEGATION = proof.fixtureDelegation ? 'true' : 'false';
+  env.FACTORY_PROOF_PROFILE = proof.profile;
+  env.FACTORY_USE_FIXTURE_DELEGATION = proof.fixtureDelegation ? 'true' : 'false';
   if (proof.profile === 'live') {
-    next.FF_REAL_SPECIALIST_DELEGATION = 'true';
-    next.SPECIALIST_DELEGATION_RUNNER = proof.runner || OPENCLAW_SPECIALIST_RUNNER;
-    if (proof.openclawBaseUrl) {
-      next.OPENCLAW_BASE_URL = proof.openclawBaseUrl;
-    }
+    env.FF_REAL_SPECIALIST_DELEGATION = 'true';
+    env.SPECIALIST_DELEGATION_RUNNER = proof.runner || OPENCLAW_SPECIALIST_RUNNER;
+    if (proof.openclawBaseUrl) env.OPENCLAW_BASE_URL = proof.openclawBaseUrl;
   } else if (proof.profile === 'fixture') {
-    next.SPECIALIST_DELEGATION_RUNNER = proof.runner || FIXTURE_SPECIALIST_RUNNER;
+    env.SPECIALIST_DELEGATION_RUNNER = proof.runner || FIXTURE_SPECIALIST_RUNNER;
   }
-  return next;
+  return env;
 }
 
 async function applyPrimaryFactoryProofProfile(options = {}) {
   const proof = await resolveFactoryProofProfile(options);
   applyFactoryProofProfileToEnv(proof, options.env || process.env);
-  if (proof.warning && options.emitWarning !== false) {
-    process.stderr.write(`${proof.warning}\n`);
-  }
+  if (proof.warning && options.emitWarning !== false) process.stderr.write(`${proof.warning}\n`);
   return proof;
 }
 
@@ -317,6 +193,18 @@ function collectAgentSessionEvidence(factoryEvidence = {}) {
   return extractSessionCandidates(factoryEvidence, []);
 }
 
+function partitionSessions(sessions) {
+  const liveSessions = sessions.filter((entry) => (
+    entry.sessionId
+    && !isFixtureSessionId(entry.sessionId)
+    && !isFixtureOwnership(entry.ownership || {})
+  ));
+  const fixtureSessions = sessions.filter((entry) => (
+    isFixtureSessionId(entry.sessionId) || isFixtureOwnership(entry.ownership || {})
+  ));
+  return { liveSessions, fixtureSessions };
+}
+
 function validateLiveSessionEvidence({
   factoryEvidence = {},
   profile = 'live',
@@ -342,9 +230,7 @@ function validateLiveSessionEvidence({
   }
 
   const sessions = collectAgentSessionEvidence(factoryEvidence);
-  const liveSessions = sessions.filter((entry) => entry.sessionId && !isFixtureSessionId(entry.sessionId) && !isFixtureOwnership(entry.ownership || {}));
-  const fixtureSessions = sessions.filter((entry) => isFixtureSessionId(entry.sessionId) || isFixtureOwnership(entry.ownership || {}));
-
+  const { liveSessions, fixtureSessions } = partitionSessions(sessions);
   if (fixtureSessions.length > 0) {
     errors.push({
       code: FACTORY_PROOF_ERROR_CODES.FIXTURE_ATTRIBUTION,
@@ -352,7 +238,6 @@ function validateLiveSessionEvidence({
       sessions: fixtureSessions.map((s) => s.sessionId),
     });
   }
-
   if (requireAtLeastOneSession && liveSessions.length === 0) {
     errors.push({
       code: FACTORY_PROOF_ERROR_CODES.MISSING_SESSION,
@@ -411,6 +296,61 @@ function attachProofMetadata(target = {}, proof = {}, validation = null) {
   };
 }
 
+function loadFactoryEvidenceFromResult(evidence = {}) {
+  let factoryEvidence = evidence.factoryEvidence || evidence.factory?.factoryEvidence || {};
+  const factoryEvidencePath = evidence.artifacts?.factoryEvidence;
+  if ((!factoryEvidence || !Object.keys(factoryEvidence).length) && factoryEvidencePath && fs.existsSync(factoryEvidencePath)) {
+    factoryEvidence = JSON.parse(fs.readFileSync(factoryEvidencePath, 'utf8'));
+  }
+  return factoryEvidence;
+}
+
+function finalizeLiveProofEvidence(evidence, proof, env = process.env) {
+  const factoryEvidence = loadFactoryEvidenceFromResult(evidence);
+  const validation = validateLiveSessionEvidence({
+    factoryEvidence,
+    profile: proof.profile,
+    runner: env.SPECIALIST_DELEGATION_RUNNER,
+    requireAtLeastOneSession: proof.profile === 'live',
+  });
+  Object.assign(evidence, attachProofMetadata(evidence, proof, validation));
+  if (proof.profile === 'live') {
+    evidence.summary = evidence.summary || { passed: false, checks: [] };
+    evidence.summary.checks = evidence.summary.checks || [];
+    evidence.summary.checks.push({
+      name: 'live_session_evidence',
+      ok: validation.ok,
+      liveSessionCount: validation.liveSessions?.length || 0,
+      errors: validation.errors,
+    });
+    evidence.summary.passed = evidence.summary.checks.every((check) => check.ok);
+    if (!validation.ok) assertLiveProofAllowsCompletion(proof, validation);
+  }
+  return { evidence, validation };
+}
+
+function writeMilestoneVerifyReport({
+  evidence,
+  proof,
+  milestone,
+  title,
+  outputDir,
+}) {
+  process.stdout.write(`${JSON.stringify({
+    ok: evidence.summary.passed,
+    milestone,
+    title,
+    outputDir,
+    proofProfile: proof.profile,
+    fixtureDelegation: proof.fixtureDelegation,
+    openclawBaseUrl: proof.openclawBaseUrl,
+    summary: evidence.summary,
+    artifacts: evidence.artifacts,
+    milestoneDComplete: evidence.artifacts?.milestoneDComplete || undefined,
+  }, null, 2)}\n`);
+  if (!evidence.summary.passed) process.exitCode = 1;
+}
+
 module.exports = {
   DEFAULT_LIVE_OPENCLAW_URL,
   DEFAULT_PROBE_TIMEOUT_MS,
@@ -435,4 +375,7 @@ module.exports = {
   validateLiveSessionEvidence,
   assertLiveProofAllowsCompletion,
   attachProofMetadata,
+  loadFactoryEvidenceFromResult,
+  finalizeLiveProofEvidence,
+  writeMilestoneVerifyReport,
 };
